requirements.txt
playwright>=1.49.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.1
click>=8.1.0
rich>=13.9.0
schedule>=1.2.2
